remove fuunctions to more meaninful locations
